-refactor reusable components -added redux store
