+++ src/pages/model/components/RegionForm/CentralValleyForm/index.jsx
@@ -1,8 +1,8 @@
 import React, { useState, useEffect } from 'react';
 import { Form, Button, Spin, Alert } from 'antd';
 import { getCentralValley } from '@/services/region';
-import styles from '../../index.less';
-import Map from '../../../../../../../components/Maps/FormMap';
+import styles from '../index.less';
+import Map from '../../../../../components/Maps/FormMap';
 
 const style = {
   labelCol: {
