+++ src/pages/model/components/RegionForm/FarmForm/index.jsx
@@ -2,8 +2,8 @@ import React, { useState, useEffect } from 'react';
 import { Form, Select, Button, Spin } from 'antd';
 import { getCVHMFarms } from '@/services/region';
 import { connect } from 'react-redux';
-import styles from '../../index.less';
-import Map from '../../../../../../../components/Maps/FormMap';
+import styles from '../index.less';
+import Map from '../../../../../components/Maps/FormMap';
 
 const { Option } = Select;
 const style = {
